GpsTime function in Flux and IFlux now returns a double instead of an int.

+++ FluxSvc/FluxSvc/IFlux.h
@@ -72,7 +72,7 @@ public:
        
     virtual void addFactory(std::string name, const ISpectrumFactory* factory )=0;
 
-    virtual int gpsTime()const=0;
+    virtual /*int*/double gpsTime()const=0;
 
 #if 0
     // get a description of the parameters that can be modified, and reference to a list of them
